Move cookie-functions.js into components folder
The cookie-functions.js file was being cloned into the deploy/javascripts folder unnecessarily
