+++ src/javascripts/components/cookie-banner.js
@@ -1,4 +1,4 @@
-import * as CookieFunctions from '../cookie-functions.js'
+import * as CookieFunctions from './cookie-functions.js'
 import 'govuk-frontend/govuk/vendor/polyfills/Function/prototype/bind'
 import 'govuk-frontend/govuk/vendor/polyfills/Event'
 import common from 'govuk-frontend/govuk/common'
